added results page and fixed caps in file names
