+++ recipE-Z/src/screens/ResultsScreen.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+import { Button, StyleSheet, Text, TextInput, View, SafeAreaView, ScrollView, Keyboard, TouchableOpacity } from 'react-native';
+import {NavigationContainer} from '@react-navigation/native';
+import {createStackNavigator} from '@react-navigation/stack';
+import defaultStyles from './stylesheet';
+
+class ResultsScreen extends React.Component {
+  render() {
+    const {navigation} = this.props;
+    return (
+      <View style={[defaultStyles.container, {backgroundColor:'#ffcece'}]}>
+      </View>
+   ); //end return
+
+ }//end render
+}//end class
+
+export default ResultsScreen;
+
+const styles = StyleSheet.create({
+});
